Create run-alpha-backup.sh
This is (as of commit time) the current script I am using to run the backup.  It is set up in the Unraid userscripts plugin to run on a custom daily schedule.  The beta script is the same, with s/alpha/beta/,  This creates a new dated log file for each job run, and moves and zips logs older than 15 days to an archive subfolder and gzips them.  I may go back later after I see how much space this all takes up and put in a 90 day or 1 year permanent log delete, but don't see the need right now.  I like having the du-xx.log that shows the current backup space usage and free space on the array at a glance, and the copied-files log should show only items which were actually copied during the run.  Credit to ChatGPT for making whipping this into shape very easy.